Style: Add menu background
