+++ src/App.css
@@ -8,3 +8,9 @@
   outline: none;
   border-bottom: 2px solid #22c55e;
 }
+
+.menuColor {
+  background-color: rgba(0, 0, 0, 0.3);
+  margin-top: 3rem;
+  height: 100vh;
+}
