+++ src/components/index.ts
@@ -3,3 +3,4 @@ export * from './About';
 export * from './Projects';
 export * from './Posts';
 export * from './Message';
+export * from './Navbar';
